Adding functions for post
