+++ src/functionality.js
@@ -1,3 +1,6 @@
+import getResponse from './addPost.js';
+import getData from './getData.js';
+
 const form = document.querySelector('.Add-form');
 
 const displayItem = (score) => {
@@ -7,19 +10,6 @@ const displayItem = (score) => {
       `;
   listPa.appendChild(listEl);
 };
-async function getResponse(scoreInfo) {
-  const response = await fetch(
-    'https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/nOWB0WyD6z5bEYJ0J5QV/scores/',
-    {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(scoreInfo),
-    }
-  );
-  if (!response.ok) {
-    throw new Error(`HTTP error! status: ${response.status}`);
-  }
-}
 form.addEventListener('submit', (e) => {
   e.preventDefault();
   const nameEl = document.querySelector('#name');
@@ -38,17 +28,7 @@ form.addEventListener('submit', (e) => {
 
 document.querySelector('.refresh-btn').addEventListener('click', () => {
   document.querySelector('.ulEl').innerHTML = ' ';
-  async function getData() {
-    const response = await fetch(
-      'https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/nOWB0WyD6z5bEYJ0J5QV/scores/'
-    );
-    if (!response.ok) {
-      throw new Error(`HTTP error! status: ${response.status}`);
-    }
-    const data = await response.json();
-    data.result.forEach((score) => {
-      displayItem(score);
-    });
-  }
   getData();
 });
+
+export default displayItem;
